created new class HarDataEntry for code separation
